update the logic for send recieve to focus on amount first
